+++ core/src/main/java/org/gbif/vocabulary/persistence/handlers/ValueListByLanguageMapTypeHandler.java
@@ -1,6 +1,6 @@
 package org.gbif.vocabulary.persistence.handlers;
 
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.io.IOException;
 import java.sql.CallableStatement;
