+++ core/src/test/java/org/gbif/vocabulary/persistence/mappers/BaseMapperTest.java
@@ -2,7 +2,7 @@ package org.gbif.vocabulary.persistence.mappers;
 
 import org.gbif.api.model.registry.LenientEquals;
 import org.gbif.vocabulary.model.VocabularyEntity;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
