+++ core/src/test/java/org/gbif/vocabulary/persistence/mappers/VocabularyMapperTest.java
@@ -3,7 +3,7 @@ package org.gbif.vocabulary.persistence.mappers;
 import org.gbif.vocabulary.PostgresDBExtension;
 import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.search.KeyNameResult;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 import org.gbif.vocabulary.persistence.parameters.NormalizedValuesParam;
 
 import java.net.URI;
