+++ core/src/test/java/org/gbif/vocabulary/service/ConceptServiceIT.java
@@ -4,7 +4,7 @@ import org.gbif.vocabulary.PostgresDBExtension;
 import org.gbif.vocabulary.model.Concept;
 import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.search.ConceptSearchParams;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 import org.gbif.vocabulary.persistence.mappers.VocabularyMapper;
 
 import java.sql.Connection;
