+++ model/src/main/java/org/gbif/vocabulary/model/Concept.java
@@ -1,7 +1,7 @@
 package org.gbif.vocabulary.model;
 
 import org.gbif.api.model.registry.LenientEquals;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.net.URI;
 import java.util.*;
