+++ model/src/main/java/org/gbif/vocabulary/model/enums/LanguageRegion.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gbif.vocabulary.model.vocabulary;
+package org.gbif.vocabulary.model.enums;
 
 import java.io.IOException;
