+++ model/src/test/java/org/gbif/vocabulary/model/VocabularyTest.java
@@ -1,6 +1,6 @@
 package org.gbif.vocabulary.model;
 
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.net.URI;
 import java.time.LocalDateTime;
