+++ vocabulary-lookup/src/main/java/org/gbif/vocabulary/lookup/VocabularyLookup.java
@@ -5,7 +5,7 @@ import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.export.ExportMetadata;
 import org.gbif.vocabulary.model.export.VocabularyExport;
 import org.gbif.vocabulary.model.normalizers.StringNormalizer;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.io.IOException;
 import java.io.InputStream;
