+++ vocabulary-rest-ws/src/main/java/org/gbif/vocabulary/restws/resources/VocabularyLanguageResource.java
@@ -1,6 +1,6 @@
 package org.gbif.vocabulary.restws.resources;
 
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 import org.gbif.vocabulary.restws.utils.Constants;
 
 import org.springframework.web.bind.annotation.GetMapping;
