+++ vocabulary-rest-ws/src/test/java/org/gbif/vocabulary/restws/resources/VocabularyResourceIT.java
@@ -3,7 +3,7 @@ package org.gbif.vocabulary.restws.resources;
 import org.gbif.vocabulary.model.Concept;
 import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.export.VocabularyExport;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 
 import java.io.IOException;
 import java.util.Arrays;
