+++ vocabulary-rest-ws/src/test/java/org/gbif/vocabulary/restws/resources/mock/ConceptResourceTest.java
@@ -6,7 +6,7 @@ import org.gbif.vocabulary.model.Concept;
 import org.gbif.vocabulary.model.Vocabulary;
 import org.gbif.vocabulary.model.search.ConceptSearchParams;
 import org.gbif.vocabulary.model.search.KeyNameResult;
-import org.gbif.vocabulary.model.vocabulary.LanguageRegion;
+import org.gbif.vocabulary.model.enums.LanguageRegion;
 import org.gbif.vocabulary.restws.model.DeprecateAction;
 import org.gbif.vocabulary.restws.model.DeprecateConceptAction;
 import org.gbif.vocabulary.restws.resources.ConceptResource;
